+++ src/features/app/DefaultCatchBoundary.tsx
@@ -6,7 +6,7 @@ import {
 	useRouter,
 } from "@tanstack/react-router";
 import type { ErrorComponentProps } from "@tanstack/react-router";
-import { Button } from "./ui/button";
+import { Button } from "../../components/ui/button";
 
 export function DefaultCatchBoundary({ error }: ErrorComponentProps) {
 	const router = useRouter();
