+++ src/features/form/form.ts
@@ -1,11 +1,6 @@
 import { createFormHook } from "@tanstack/react-form";
 
-import {
-	Select,
-	SubscribeButton,
-	TextArea,
-	TextField,
-} from "../components/FormComponents";
+import { Select, SubscribeButton, TextArea, TextField } from "./FormComponents";
 import { fieldContext, formContext } from "./form-context";
 
 export const { useAppForm } = createFormHook({
