+++ src/features/profile/ProfileEdit.tsx
@@ -1,11 +1,15 @@
 import { useUser } from "@clerk/clerk-react";
 import { api } from "@convex-server/_generated/api";
-import { useNavigate } from "@tanstack/react-router";
-import { useMutation } from "convex/react";
+import type { Id } from "@convex-server/_generated/dataModel";
+import { useMutation, useQuery } from "convex/react";
+import { useEffect } from "react";
 import { toast } from "sonner";
 import { z } from "zod";
-import { useAppForm } from "../../hooks/form";
-import { Avatar, AvatarFallback, AvatarImage } from "../ui/avatar";
+import {
+	Avatar,
+	AvatarFallback,
+	AvatarImage,
+} from "../../components/ui/avatar";
 import {
 	Card,
 	CardContent,
@@ -13,7 +17,8 @@ import {
 	CardFooter,
 	CardHeader,
 	CardTitle,
-} from "../ui/card";
+} from "../../components/ui/card";
+import { useAppForm } from "../form/form";
 
 const schema = z.object({
 	name: z.string().min(1, "Full name is required"),
@@ -22,44 +27,45 @@ const schema = z.object({
 	avatarUrl: z.string(),
 });
 
-export function ProfileSetup() {
+export function ProfileEdit() {
 	const { user } = useUser();
-	const navigate = useNavigate();
-	const createProfile = useMutation(api.users.createProfile);
+	const profile = useQuery(api.users.getProfile, { clerkId: user?.id || "" });
+	const updateProfile = useMutation(api.users.updateProfile);
 
 	const form = useAppForm({
 		defaultValues: {
-			name: user?.fullName || "",
-			email: user?.primaryEmailAddress?.emailAddress || "",
-			bio: "",
-			avatarUrl: user?.imageUrl || "",
+			name: profile?.name || "",
+			email: profile?.email || "",
+			bio: profile?.bio || "",
+			avatarUrl: profile?.avatarUrl || "",
 		},
 		validators: {
 			onBlur: schema,
 		},
 		onSubmit: async ({ value }) => {
 			try {
-				await createProfile({
+				await updateProfile({
+					id: profile?._id as Id<"users">,
 					...value,
-					clerkId: user?.id || "",
 				});
-				toast.success("Profile created successfully!");
-				navigate({ to: "/dashboard" });
+				toast.success("Profile updated successfully!");
 			} catch (error) {
-				toast.error("Failed to create profile. Please try again.");
-				console.error("Profile creation error:", error);
+				toast.error("Failed to update profile. Please try again.");
+				console.error("Profile update error:", error);
 			}
 		},
 	});
+
+	if (!profile) {
+		return <div>Loading...</div>;
+	}
 
 	return (
 		<div className="container mx-auto max-w-2xl p-4">
 			<Card>
 				<CardHeader>
-					<CardTitle>Complete Your Profile</CardTitle>
-					<CardDescription>
-						Please provide some information about yourself to get started.
-					</CardDescription>
+					<CardTitle>Edit Profile</CardTitle>
+					<CardDescription>Update your profile information.</CardDescription>
 				</CardHeader>
 				<form
 					onSubmit={(e) => {
@@ -112,7 +118,7 @@ export function ProfileSetup() {
 					</CardContent>
 					<CardFooter className="flex justify-end space-x-2">
 						<form.AppForm>
-							<form.SubscribeButton label="Complete Setup" />
+							<form.SubscribeButton label="Save Changes" />
 						</form.AppForm>
 					</CardFooter>
 				</form>
