+++ src/features/profile/profile-queries.ts
@@ -1,0 +1,3 @@
+import { convexQuery, useConvexMutation } from "@convex-dev/react-query";
+import { api } from "@convex-server/_generated/api";
+import { useMutation } from "@tanstack/react-query";
