+++ src/features/trello-board/NewColumn.tsx
@@ -1,10 +1,10 @@
 import { useRef, useState } from "react";
 import invariant from "tiny-invariant";
 
-import { CancelButton } from "@/components/CancelButton";
-import { SaveButton } from "@/components/SaveButton";
-import { Icon } from "../icons/icons";
-import { useCreateColumnMutation } from "../queries";
+import { CancelButton } from "@/features/trello-board/CancelButton";
+import { SaveButton } from "@/features/trello-board/SaveButton";
+import { Icon } from "../../icons/icons";
+import { useCreateColumnMutation } from "../../queries";
 
 export function NewColumn({
 	boardId,
