+++ src/integrations/tanchat/header-user.tsx
@@ -1,4 +1,4 @@
-import AIAssistant from "../../components/example-AIAssistant";
+import AIAssistant from "../../features/examples/example-AIAssistant";
 
 export default function HeaderAIAssistant() {
 	return <AIAssistant />;
