+++ src/router.tsx
@@ -6,8 +6,8 @@ import { createRouter as createTanStackRouter } from "@tanstack/react-router";
 import { routerWithQueryClient } from "@tanstack/react-router-with-query";
 import { createIsomorphicFn } from "@tanstack/react-start";
 import { ConvexProvider, ConvexReactClient } from "convex/react";
-import { DefaultCatchBoundary } from "./components/DefaultCatchBoundary";
 import { NotFound } from "./components/NotFound";
+import { DefaultCatchBoundary } from "./features/app/DefaultCatchBoundary";
 import { routeTree } from "./routeTree.gen";
 
 export function createRouter() {
