+++ src/routes/demo.form.address.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute } from "@tanstack/react-router";
 
-import { useAppForm } from "../hooks/form";
+import { useAppForm } from "../features/form/form";
 
 export const Route = createFileRoute("/demo/form/address")({
 	component: AddressForm,
