+++ src/routes/demo.form.simple.tsx
@@ -1,7 +1,7 @@
 import { createFileRoute } from "@tanstack/react-router";
 import { z } from "zod";
 
-import { useAppForm } from "../hooks/form";
+import { useAppForm } from "../features/form/form";
 
 export const Route = createFileRoute("/demo/form/simple")({
 	component: SimpleForm,
